refactor: merge exceptions files
