+++ aiarena/core/services/__init__.py
@@ -2,7 +2,6 @@
 from .bot_statistics import BotStatistics
 from .bots import Bots
 from .competitions import Competitions
-from .exceptions import MatchRequestException
 from .ladders import Ladders
 from .match_requests import MatchRequests
 from .matches import Matches
