--- aiarena/core/services/exceptions.py
@@ -1,3 +1,0 @@
-class MatchRequestException(Exception):
-    def __init__(self, message):
-        super().__init__(message)
