+++ aiarena/core/services/internal/match_requests.py
@@ -4,7 +4,7 @@ from constance import config
 
 from aiarena.core.models import Map, Match, Result
 
-from ..exceptions import MatchRequestException
+from ...exceptions import MatchRequestException
 from ..supporter_benefits import SupporterBenefits
 from .maps import Maps
 from .matches import create
